requirements.txt
Flask==2.3.3
Pillow==10.0.1
exifread==3.0.0
flask-cors==4.0.0
werkzeug==2.3.7
